Switch S2I builds to --as-dockerfile
Adjust `s2i_inner`, the main method used for invoking the s2i binary,
such that it uses the "--as-dockerfile" build parameter. This causes
it to generate the image sources required to perform the S2I build,
but not to perform the build itself. There are two reasons to do
this:

1. It's closer to what OpenShift's internal source-to-image process
   does. We've hit some bugs which are triggered in an OpenShift
   deployment that are not being exercised by s2i in standalone
   mode (see some of the discussion in e.g. OPENJDK-2824).

2. Doing the build stage ourselves allows us to vary parameters of
   the build -- such as the container tooling to use -- in the
   future, smoothing progress towards running tests with Podman.

One known caveat: using s2i end-to-end, the scripts-url value is
determined in the following priority order

1. --scripts-url argument passed to s2i
2. application source .s2i/bin directory
3. path from builder image's io.openshift.s2i.scripts-url label

This commit currently maps (3) to (1), so in a scenario where the
label is specified in the builder image, and the application source
contains a .s2i/bin binary, before, the application copy would take
precedence, but now it will be ignored. This will break some tests.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>